fix: local storage settings hydration
