+++ packages/core/src/lib/game-manager/game/game.ts
@@ -1,5 +1,5 @@
 import { IGame } from '../../interfaces/game.interface';
-import { DifficultyStrategy } from '../../settings-manager/modules/difficulty-strategy/difficulty.strategy';
+import { DifficultyStrategy } from '../../settings-manager/default-settings-handler/default-settings.handler';
 import {
   NoCurrentReactionError,
   NoPreviousReactionError,
