+++ packages/core/src/lib/settings-manager/default-settings-handler/decorators/rehydrate.middleware.ts
@@ -1,9 +1,11 @@
 import { Settings } from '../../../interfaces';
 import {
   EasyDifficultyStrategy,
+  Middleware,
+  TimerOnGuessDifficulty,
   UnlimitedLivesBut5050ChanceOfGameOverDifficulty,
-} from '../../modules';
-import { Middleware } from '../default-settings.handler';
+  VariableDeviationDifficulty,
+} from '../default-settings.handler';
 
 export const RehydrateSettingsMiddleware: Middleware<Settings> = (
   context,
@@ -12,14 +14,19 @@ export const RehydrateSettingsMiddleware: Middleware<Settings> = (
   const hasDifficultyStrategy =
     context.difficulty.key === 'DIFFICULTY_STRATEGY';
   if (hasDifficultyStrategy) {
-    if (context.difficulty.id === 'EASY_DIFFICULTY') {
-      context.difficulty = new EasyDifficultyStrategy();
-    } else if (
-      context.difficulty.id ===
-      'UNLIMITED_LIVES_BUT_50_50_CHANCE_OF_GAME_OVER_DIFFICULTY'
-    ) {
-      context.difficulty =
-        new UnlimitedLivesBut5050ChanceOfGameOverDifficulty();
+    switch (context.difficulty.id) {
+      case 'EASY_DIFFICULTY':
+        context.difficulty = new EasyDifficultyStrategy();
+        break;
+      case 'UNLIMITED_LIVES_BUT_50_50_CHANCE_OF_GAME_OVER_DIFFICULTY':
+        context.difficulty =
+          new UnlimitedLivesBut5050ChanceOfGameOverDifficulty();
+        break;
+      case 'TIMER_ON_GUESS_DIFFICULTY':
+        context.difficulty = new TimerOnGuessDifficulty();
+        break;
+      case 'VARIABLE_DEVIATION_DIFFICULTY':
+        context.difficulty = new VariableDeviationDifficulty();
     }
   }
   return next();
