--- packages/core/src/lib/settings-manager/modules/difficulty-strategy/index.ts
@@ -1,1 +1,0 @@
-export * from './difficulty.strategy';
