+++ packages/core/src/lib/settings-manager/settings-manager.ts
@@ -5,8 +5,10 @@ import {
 } from '../game-manager';
 import { Coloring, Settings } from '../interfaces';
 import { ObserverSubject } from '../observer';
-import { DefaultSettingsHandlerImpl } from './default-settings-handler';
-import { DifficultyStrategy } from './modules';
+import {
+  DefaultSettingsHandlerImpl,
+  DifficultyStrategy,
+} from './default-settings-handler';
 
 export enum SettingsManagerEvent {
   SET_COLORING = 'SET_COLORING',
